% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_dgirt.r
\name{apply_dgirt}
\alias{apply_dgirt}
\title{Summarize dgirt samples}
\usage{
apply_dgirt(dgirt_output, dgirt_input, fun = mean.default)
}
\arguments{
\item{dgirt_output}{Return value of `dgirt`, a `stanfit` object.}

\item{dgirt_input}{Return value of `wrangle`.}

\item{fun}{A single scalar function like `mean`.}
}
\value{
A list of tables summarizing the posterior distribution of each model parameter.
}
\description{
`apply_dgirt` applies a scalar function over dgirt sampler iterations for each parameter that appears in the `stanfit`
object returned by `dgirt`.
}

