% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.r
\name{create_weights}
\alias{create_weights}
\title{Create individual survey weights from population targets}
\usage{
create_weights(level1, arg)
}
\arguments{
\item{level1}{\code{data.frame} with survey responses at the lowest level of aggregation.}

\item{arg}{Arguments to `wrangle`}
}
\value{
\code{data_frame} of survey responses with new weights normalized to have mean 1 in each period.
}
\description{
Strata in the sampled data must be a subset of those in the population targets.
There can't be missingness in the stratifying variables.
}

