% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dgirtfit.r, R/methods-dgirtfit.r
\docType{class}
\name{dgirtFit-class}
\alias{dgirtFit}
\alias{dgirtFit-class}
\alias{dgirtfit}
\alias{dgirtfit-class}
\alias{extract,dgirtFit-method}
\alias{get_posterior_mean,dgirtFit-method}
\alias{summary,dgirtFit-method}
\title{Fitted DGIRT Model.}
\usage{
\S4method{summary}{dgirtFit}(object = dgirtFit, ...)

\S4method{extract}{dgirtFit}(object = dgirtFit, ...)

\S4method{get_posterior_mean}{dgirtFit}(object = dgirtFit, ...)
}
\description{
All `rstan` methods for superclass `\link[rstan]{stanfit}` are available.
Descriptive labels for parameters on time periods, local geographic areas,
and grouping variables will be added to most output.
}
\section{Slots}{

\describe{
\item{\code{dgirt_in}}{Data used to fit the model.}
}}
\examples{
data(toy_dgirtfit)
# summarize the fitted results
summary(toy_dgirtfit, pars = 'xi')

# get posterior means with a convenience function
get_posterior_mean(toy_dgirtfit, pars = 'xi')

# generally apply functions to posterior samples after warmup; n.b.
# `as.array` is iterations x chains x parameters so `MARGIN = 3` applies
# `FUN` over iterations and chains
apply(as.array(toy_dgirtfit, pars = 'xi'), 3, mean)

# access the posterior samples
as.array(toy_dgirtfit, pars = 'xi')  # samples after warmup
as.data.frame(toy_dgirtfit, pars = 'xi')  # all samples
extract(toy_dgirtfit, pars = 'theta_bar')  # all samples
}
\seealso{
`\link[rstan]{stanfit-class}`
}

