% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgirt.r
\name{dgirt}
\alias{dgirt}
\title{Estimate a DGIRT model}
\usage{
dgirt(shaped_data, ...)
}
\arguments{
\item{shaped_data}{Output from `shape`.}

\item{...}{Further arguments to `\link[rstan]{stan}`.}
}
\value{
A `dgirtfit` object that extends `\link[rstan]{stanfit-class}`. For
details see `\link{dgirtfit-class}`.
}
\description{
`dgirt` makes a call to `\link[rstan]{stan}` with the Stan code and data for
a DGIRT model. By default the call overrides the `stan` defaults for `pars`
to specify typical DGIRT parameters of interest and sets `iter_r` to `1L`.
}

