% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dgmrp_fit.r
\docType{class}
\name{dgmrp_fit-class}
\alias{dgmrp_fit-class}
\alias{dgmrp_fit}
\title{A class for fitted dynamic group MRP models}
\description{
\code{\link{dgmrp}} returns a fitted model object of class \code{dgmrp_fit},
which inherits from \code{\link{dgo_fit}}.
}
\section{Slots}{

\describe{
\item{\code{dgirt_in}}{\code{\link{dgirtin-class}} data used to fit the model.}
}}

\seealso{
\code{\link{dgirt_fit}} \code{\link{dgo_fit}}
}
