% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{dgo}
\alias{dgo}
\alias{dgo-package}
\title{dgo: Dynamic Estimation of Group-level Opinion}
\description{
Fit dynamic group-level IRT models from individual or aggregated item
response data. This package handles common preprocessing tasks and extends
functions for inspecting results, poststratification, and quick iteration
over alternative models.
}
