% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dgo_fit.r
\docType{class}
\name{dgo_fit-class}
\alias{dgo_fit-class}
\alias{dgo_fit}
\title{Class \code{dgo_fit}: a class for fitted models}
\description{
\code{dgo_fit} is a superclass for \code{\link{dgirt_fit}} and
\code{\link{dgmrp_fit}} that inherits from the
\code{\link[rstan]{stanfit-class}} in the \code{\link{rstan}} package.
}
\section{Slots}{

\describe{
\item{\code{dgirt_in}}{\code{\link{dgirtin-class}} data used to fit the model.}

\item{\code{call}}{The function call that returned the \code{dgo_fit} object.}
}}

\seealso{
\code{\link{dgmrp_fit}} \code{\link{dgo_fit}}
}
