% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.r
\name{format_data}
\alias{format_data}
\title{Format data for use with dgirt}
\usage{
format_data(data = list(level1, level2 = NULL, targets = NULL),
  vars = list(items, groups, time_id, geo_id, survey_id, survey_weight,
  preweights = NULL, preweight_proportion = NULL, level2_modifiers = NULL,
  level2_period1_modifiers = NULL), filters = list(periods = NULL, geo_ids =
  NULL, min_surveys = 1L, min_periods = 1L), params = list(separate_periods =
  FALSE, difficulty_count = 1L, constant_item = TRUE, delta_tbar_prior_mean =
  0.5, delta_tbar_prior_sd = 0.5, innov_sd_delta_scale = 2.5,
  innov_sd_theta_scale = 2.5))
}
\arguments{
\item{data}{Named `list` of `data.frame`s containing:
\describe{
  \item{level1}{Survey responses at the lowest level of aggregation (e.g. individual).}
  \item{level2}{Survey responses at higher level
     of aggregation. Currently restricted to geographic units, and must be
     accompanied by `level2_modifiers`.}
  \item{targets}{Population targets for stratification. Strata in
     `level1` must be a subset of those in the target data,
     and there can't be missingness in the stratifying variables.}
}}

\item{vars}{Named `list` of character vectors that give the variable or variables in representing:
\describe{
  \item{items}{Item responses in `data$level1`}
  \item{groups}{Respondent characteristics in `data$level1`.}
  \item{time_id}{Time period in `data$level1` and `data$level2`.}
  \item{geo_id}{Geographic identifier in `data$level1` and `data$level2`.}
  \item{survey_id}{Survey identifier in `data$level1`.}
  \item{survey_weight}{Survey weight in `data$level1`.}
  \item{target_groups}{Respondent characteristics in `data$targets`.}
  \item{target_proportion}{Population proportions in `data$targets`.}
  \item{level2_modifiers}{Geographic characteristics in `data$level2`.}
  \item{level2_period1_modifiers}{Geographic characteristics in `data$level2` to be used in place of `vars$level2_modifiers` in the first period.}
}}

\item{filters}{Named `list` of filters that will be applied to the data.
\describe{
  \item{periods}{A character vector of time periods to be used in estimation,
     including unobserved periods or excluding observed periods as desired.
     Defaults to the values of the `vars$time_id` variable in the data.}
  \item{geo_ids}{A character vector giving the subset of values of `vars$geo_id` in the data to be used in estimation.
     Defaults to the values of `vars$geo_id` in the data.}
  \item{min_surveys}{A positive integer giving the minimum surveys in which an item must appear to be used in estimation.}
  \item{min_periods}{A positive integer giving the minimum periods in which an item must appear to be used in esimation.}
}}

\item{params}{Named `list` of modeling choices.
\describe{
  \item{separate_periods}{Logical for whether estimates should be pooled over time. No pooling if `TRUE`.}
  \item{constant_item}{Logical for whether item parameters should be constant over time.}
  \item{difficulty_count}{Number of difficulty parameters.}
  \item{delta_tbar_prior_mean}{Prior mean of $\\bar{\\delta_t}$.}
  \item{delta_tbar_prior_sd}{Prior standard deviation of $\\bar{\\delta_t}$.}
  \item{innov_sd_delta_scale}{Prior scale of innovation parameter for standard deviation of $\\bar{\\delta_t}$.}
  \item{innov_sd_theta_scale}{Prior scale of innovation for standard deviation of group ability parameter $\\bar{\\theta_t}$.}
}}
}
\value{
\code{list} List formatted for `run_dgirt`.
}
\description{
Format data for use with dgirt
}

