% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mgirt_fit.R
\docType{class}
\name{mgirt_fit-class}
\alias{mgirt_fit-class}
\alias{mgirt_fit}
\title{A class for fitted dynamic group multinomial IRT models}
\description{
\code{\link{mgirt}} returns a fitted model object of class \code{mgirt_fit},
which inherits from \code{\link[rstan]{stanfit-class}} in the
\code{\link{rstan}} package.
}
\section{Slots}{

\describe{
\item{\code{mgirt_in}}{\code{\link{mgirtin-class}} data used to fit the model.}

\item{\code{call}}{The function call that returned the \code{mgirt_fit} object.}
}}

