% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moc.R
\name{moc}
\alias{moc}
\title{Propagate uncertainty through the method of composition}
\usage{
moc(data, model, iter_var = "iteration", iter_values = NULL,
  vc_fun = stats::vcov)
}
\arguments{
\item{data}{A data frame containing multiple samples from the
measurement-error distribution of one or more variables.}

\item{model}{A fitted model object, from a model previously estimated (e.g.,
on the whole dataset or one sample).}

\item{iter_var}{The name of the variable in \code{data} that identifies different
samples.}

\item{iter_values}{Optionally, a vector of unique values of iteration variable
\code{iter_var}, to which \code{data} will be subset.}

\item{vc_fun}{A function for extracting the variance-covariance matrix of the
the parameters estimated by \code{model}.}
}
\value{
Samples from distribution of model parameters.
}
\description{
This function propagates uncertainty from variables measured with error by
 estimating a model on many samples of a dataset and then sampling from the
 distribution of the model parameters, a procedure known as the "method of
 composition" (Tanner 1996, 52; Treier and Jackman 2008).
}
\examples{
### Randomly Generated Data
set.seed(1)
n.obs <- 100
n.samps <- 1000
y0 <- rnorm(n.obs)
x0 <- rnorm(n.obs)
samp.df <- data.frame(iteration=rep(seq_len(n.samps), each=n.obs),
                      x=rep(x0, n.samps) + rnorm(n.obs*n.samps),
                      y=rep(y0, n.samps))
mod_rand <- lm(y ~ x, data=subset(samp.df, iteration==1))
summary(mod_rand)
moc_rand <- moc(samp.df, mod_rand)
colMeans(moc_rand) ## point estimate
apply(moc_rand, 2, sd) ## estimated standard error

\dontrun{
if (require(sandwich, quietly = TRUE)) {
  ### DGO Output 
  dgirt_in_abortion <- shape(opinion, item_names = "abortion", time_name = "year",
                             geo_name = "state", group_names = "race3", 
                             geo_filter = c("CA", "GA", "LA", "MA"),
                             id_vars = "source")
 
  dgmrp_out_abortion <- dgmrp(dgirt_in_abortion, iter = 1500, chains = 4, 
                              cores = 4, seed = 42)
  d <- as.data.frame(dgmrp_out_abortion)
 
  samples = lapply(unique(d$iteration), function(x) {
    poststratify(d[iteration == x],
                 annual_state_race_targets, 
                 strata_names = c("state", "year"),
                 aggregated_names = "race3")
  })
  samples = rbindlist(samples, idcol = "iteration")
 
  ps <- poststratify(
    dgmrp_out_abortion, 
    annual_state_race_targets, 
    strata_names = c("state", "year"), 
    aggregated_names = "race3"
  )
 
  mod_dgmrp <- lm(value ~ 0 + state, ps)  
  summary(mod_dgmrp)
  sqrt(diag(sandwich::vcovHC(mod_dgmrp)))
 
  moc_dgmrp <- moc(data=samples, model=mod_dgmrp, vc_fun=sandwich::vcovHC)
  colMeans(moc_dgmrp)
  apply(moc_dgmrp, 2, sd)
}
}
}
\references{
Tanner, Martin A. 1996. \emph{Tools for Statistical Inference Methods for the
Exploration of Posterior Distributions and Likelihood Functions.} 3rd ed. New
York: Springer.

Treier, Shawn, and Simon Jackman. 2008. "Democracy as a Latent Variable." 
\emph{American Journal of Political Science} 52 (1): 201–217.
}
\author{
Devin Caughey
}
