% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{class}
\name{Model-class}
\alias{Model-class}
\alias{Model}
\title{A class for Stan model data}
\description{
A class for Stan model data
}
\section{Slots}{

\describe{
\item{\code{D}}{Number of difficulty parameters per question.}

\item{\code{G}}{Number of covariate groups.}

\item{\code{G_hier}}{Not implemented.}

\item{\code{H}}{Number of predictors for geographic unit effects.}

\item{\code{Hprior}}{Number of predictors for geographic unit effects in first time
period.}

\item{\code{N}}{Number of covariate groups.}

\item{\code{NNl2}}{Not implemented.}

\item{\code{N_observed}}{Number of covariate groups with at least one trial.}

\item{\code{P}}{Number of hierarchical parameters, including geographic parameters.}

\item{\code{Q}}{Number of items, after dichotomizing polytomous items.}

\item{\code{S}}{Number of geographic parameters.}

\item{\code{SSl2}}{Not implemented.}

\item{\code{T}}{Number of time periods.}

\item{\code{WT}}{Weight array.}

\item{\code{XX}}{Design matrix for hierarchical variables.}

\item{\code{ZZ}}{Data for geographic model.}

\item{\code{ZZ_prior}}{Data for geographic model in the first time period.}

\item{\code{constant_item}}{Flag for whether item parameters are constant.}

\item{\code{delta_tbar_prior_mean}}{Prior mean for \code{delta_tbar}.}

\item{\code{delta_tbar_prior_sd}}{Prior standard deviation for \code{delta_tbar}.}

\item{\code{hierarchical_model}}{Flag for whether to use the hierarchical model.}

\item{\code{innov_sd_delta_scale}}{Scale parameter for innovation of \code{sd_delta}.}

\item{\code{innov_sd_theta_scale}}{Scale parameter for innovation of \code{sd_theta}.}

\item{\code{l2_only}}{Not implemented.}

\item{\code{n_vec}}{Vector of covariate group trial counts.}

\item{\code{observed}}{Non-missing indexes for \code{n_vec} and \code{s_vec}.}

\item{\code{s_vec}}{Vector of covariate group success counts.}

\item{\code{separate_t}}{Flag for whether to smooth estimates over time periods.}
}}

