% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modgirt.R
\name{modgirt}
\alias{modgirt}
\title{Fit a multinomial group IRT model}
\usage{
modgirt(shaped_data, version = "2018_08_18_mdgirt_ord", model = NULL,
  evolving_alpha = FALSE, ...)
}
\arguments{
\item{shaped_data}{Data shaped for the model. A function for accomplishing
this is not yet implemented.}

\item{...}{Further arguments, passed to \code{\link[rstan]{stan}}.}
}
\value{
A \code{\link{modgirt_fit-class}} object that extends
\code{\link[rstan]{stanfit-class}}. (Not yet implemented.)
}
\description{
Fit a multinomial group IRT model
}
