% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-modgirt_fit.R
\docType{class}
\name{modgirt_fit-class}
\alias{modgirt_fit-class}
\alias{modgirt_fit}
\title{A class for fitted MODGIRT models}
\description{
\code{\link{modgirt}} returns a fitted model object of class \code{modgirt_fit},
which inherits from \code{\link[rstan]{stanfit-class}} in the
\code{\link{rstan}} package.
}
\section{Slots}{

\describe{
\item{\code{call}}{The function call that returned the \code{modgirt_fit} object.}

\item{\code{items}}{Names of item response variables.}

\item{\code{time}}{Name of a time variable.}

\item{\code{geo}}{Name of a geographic variable.}

\item{\code{demo}}{Names of further grouping variables, usually demographic.}
}}

