% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-modgirt_fit.R
\docType{methods}
\name{show,modgirt_fit-method}
\alias{show,modgirt_fit-method}
\alias{print,modgirt_fit-method}
\alias{print.modgirt_fit}
\alias{get_elapsed_time,modgirt_fit-method}
\title{\code{print} method for \code{modgirt_fit-class} objects}
\usage{
\S4method{show}{modgirt_fit}(object)

\S4method{print}{modgirt_fit}(x)

print.modgirt_fit(x)

\S4method{get_elapsed_time}{modgirt_fit}(object, ...)
}
\arguments{
\item{object}{A \code{modgirt_fit-class} object}

\item{x}{A \code{modgirt_fit-class} object}

\item{...}{Further arguments to \code{\link{stanfit-class}} methods.}

\item{pars}{Parameter name(s)}
}
\description{
\code{print} method for \code{modgirt_fit-class} objects

Chain run times
}
