% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-modgirt_in.R
\docType{class}
\name{modgirt_in-class}
\alias{modgirt_in-class}
\alias{modgirt_in}
\alias{summary,modgirt_in-method}
\alias{show,modgirt_in-method}
\title{A class for item response data shaped for modgirt}
\usage{
\S4method{summary}{modgirt_in}(object, ...)

\S4method{show}{modgirt_in}(object)
}
\arguments{
\item{object}{An object of class \code{Individual}}

\item{...}{Further arguments.}
}
\description{
A class for item response data shaped for modgirt
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: Methods

\item \code{show}: Methods
}}

\section{Slots}{

\describe{
\item{\code{items}}{Names of item response variables.}

\item{\code{time}}{Name of a time variable.}

\item{\code{geo}}{Name of a geographic variable.}

\item{\code{demo}}{Names of further grouping variables, usually demographic.}

\item{\code{stan_data}}{Data shaped for estimation.}
}}

