% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.r
\name{name_pars}
\alias{name_pars}
\title{Extract and name parameters}
\usage{
name_pars(dgirt_output, dgirt_input)
}
\arguments{
\item{dgirt_output}{Return value of `dgirt`, a `stanfit` object.}

\item{dgirt_input}{Return value of `wrangle`.}
}
\value{
What `rstan::extract` returns, but with dimension names.
}
\description{
`name ` is a wrapper for rstan::extract that attaches names to parameters
using the data originally passed to `dgirt`.
}

