% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-opinion.r
\docType{data}
\name{opinion}
\alias{opinion}
\title{\code{dgirt} example data: item responses}
\format{A \code{data.frame} with 78,709 observations of 10 variables.}
\usage{
opinion
}
\description{
A table of item responses given by survey respondents and their
characteristics. Data are from the Cooperative Congressional Election Study
(CCES), 2006-2008
}
\examples{
opinion
}
\seealso{
\url{http://projects.iq.harvard.edu/cces/data}
}

