% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-optimize_output.r
\docType{data}
\name{optimize_output}
\alias{optimize_output}
\title{Example of dgirt optimization output}
\format{A list whose elements correspond to DGIRT parameters and
contain estimates of those parameters.}
\usage{
optimize_output
}
\description{
An example of the return value of `dgirt` when `method = "optimize"`.
}
\examples{
optimize_output
}

