% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dgirtfit-plot.r
\docType{methods}
\name{dgirt_plot}
\alias{dgirt_plot}
\alias{plot,dgirtfit,missing-method}
\title{\code{dgirt_plot} plot \code{dgirtfit}-class objects}
\usage{
dgirt_plot(x, y_fun = "median", y_min = "q_025", y_max = "q_975",
  pars = "theta_bar")

\S4method{plot}{dgirtfit,missing}(x, y, ...)
}
\arguments{
\item{x}{A \code{dgirtfit-class} object.}

\item{y_fun}{Summary function to be plotted as \code{y}.}

\item{y_min}{Summary function giving the \code{ymin} argument for a
\code{geom_pointrange} object.}

\item{y_max}{Summary function giving the \code{ymax} argument for a
\code{geom_pointrange} object.}

\item{pars}{Selected parameter.}

\item{y}{Ignored.}

\item{...}{Further arguments to \code{\link{dgirt_plot}}.}
}
\description{
\code{dgirt_plot} plot \code{dgirtfit}-class objects

\code{plot} plot method for \code{dgirtfit}-class objects
}
\examples{
dgirt_plot(toy_dgirtfit) 
dgirt_plot(toy_dgirtfit, y_min = NULL, y_max = NULL)
p <- dgirt_plot(toy_dgirtfit)
p \%+\% ylab("posterior median")

plot(toy_dgirtfit)
}

