% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_means}
\alias{plot_means}
\title{Plot group means}
\usage{
plot_means(theta_bars, time_id, facet_var, jitter = FALSE)
}
\arguments{
\item{theta_bars}{Result of either `dgirt` if `method = "optimize"` or `extract_dgirt` if `method = "rstan"` (the default).}

\item{time_id}{Variable in theta_bars representing time, which will be used appear on the horizontal axis. A character vector.}

\item{facet_var}{Name(s) of variable(s) in theta_bars representing groups that will be plotted in separate facets. A character vector.}

\item{jitter}{Logical; should points be jittered?}
}
\description{
Plot group means
}
\examples{
data(optimized_theta_bars)
optimized_theta_bars$year <- as.integer(optimized_theta_bars$year)
plot_means(optimized_theta_bars, "year", "state", jitter = TRUE)
}

