% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratify.r
\name{poststratify}
\alias{poststratify}
\title{Poststratify Estimates.}
\usage{
poststratify(group_means, target_data, group_names, strata_names = c("year",
  "state"), prop_name = "proportion", check_sums = NULL)
}
\arguments{
\item{group_means}{A table of estimates.}
}
\value{
table of poststratified group means
}
\description{
Poststratify Estimates.
}

