% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dgirtfit.r
\docType{methods}
\name{rhats,dgirtfit-method}
\alias{rhats,dgirtfit-method}
\title{\code{rhats}: extract split R-hats from \code{dgirtfit}-class objects}
\usage{
\S4method{rhats}{dgirtfit}(x, pars = "theta_bar")
}
\arguments{
\item{x}{A \code{dgirtfit}-class object}

\item{pars}{Parameter name(s)}
}
\value{
A table giving split R-hats for model parameters
}
\description{
\code{rhats}: extract split R-hats from \code{dgirtfit}-class objects
}
\examples{
rhats(toy_dgirtfit)
}

