% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-modgirt_fit.R
\docType{methods}
\name{rhats,modgirt_fit-method}
\alias{rhats,modgirt_fit-method}
\title{Split R-hats}
\usage{
\S4method{rhats}{modgirt_fit}(x, pars = "bar_theta")
}
\value{
A table giving split R-hats for model parameters
}
\description{
Split R-hats
}
