% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dgirtfit.r
\name{rhats}
\alias{rhats}
\title{\code{rhats}: extract split R-hats from \code{dgirtfit}-class objects}
\usage{
rhats(x, ...)
}
\arguments{
\item{x}{A \code{dgirtfit}-class object}

\item{...}{Method arguments}
}
\description{
\code{rhats}: extract split R-hats from \code{dgirtfit}-class objects
}

