% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-rstan_output.r
\docType{data}
\name{rstan_output}
\alias{rstan_output}
\title{Example of dgirt rstan output}
\format{An object of class `stanfit`.}
\usage{
rstan_output
}
\description{
An example of the return value of `dgirt` when `method = "rstan"` (the default).
}
\examples{
rstan_output
}

