% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-state_demographics.r
\docType{data}
\name{state_demographics}
\alias{state_demographics}
\title{State-level demographic marginals}
\format{A \code{dplyr} \code{tbl_df} with 312,120 rows and 7 columns.}
\usage{
state_demographics
}
\description{
A dataset with the proportion of each state's population 1960-2010 in strata defined by four demographic variables.
}
\details{
\itemize{
 \item state
 \item year
 \item race
 \item female
 \item education
 \item age
}
}
\examples{
state_demographics
}

