% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-state_opinion.r
\docType{data}
\name{state_opinion}
\alias{state_opinion}
\title{State-level policy item responses}
\format{A `dplyr` `tbl_df` with 295,248 rows and 33 columns.}
\usage{
state_opinion
}
\description{
A dataset of survey item responses and characteristics of respondents.
Data are from the Cooperative Congressional Election Study (CCES) 2006-2014.
`Q_` prepends the names of item response variables.
}
\examples{
state_opinion
}

