\docType{data}
\name{state_year_targets}
\alias{state_year_targets}
\title{\code{dgirt} example data: state-year population targets}
\format{A \code{data.frame} with 459 observations of 4 variables.}
\usage{
state_year_targets
}
\description{
A table giving the population proportions of state-years by race, 2006-2008.
}
\examples{
head(state_year_targets)
}

