% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-states.r
\docType{data}
\name{states}
\alias{states}
\title{State Demographics.}
\format{A `data.frame` with 153 observations of 7 variables.}
\usage{
states
}
\description{
A table giving demographic characteristics of American states 2006-2008.
}
\examples{
states
}

