% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-toy_dgirt_in.r
\docType{data}
\name{toy_dgirt_in}
\alias{toy_dgirt_in}
\title{A Minimal Example of the dgirtIn Class.}
\format{A `dgirtIn` object.}
\usage{
toy_dgirt_in
}
\description{
`\link{shape}` returns a `dgirtIn`-class object used with `\link{dgirt}` for
DGIRT modeling. `toy_dgirt_in` is a minimal `dgirtIn` object, mostly for
use in development.
}
\examples{
toy_dgirt_in
}

