% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgirt.r
\name{dgirt}
\alias{dgirt}
\title{Estimate a DGIRT model}
\usage{
dgirt(shaped_data, separate_t = FALSE, delta_tbar_prior_mean = 0.5,
  delta_tbar_prior_sd = 0.5, innov_sd_delta_scale = 2.5,
  innov_sd_theta_scale = 2.5, ...)
}
\arguments{
\item{shaped_data}{Output from `shape`.}

\item{separate_t}{Whether smoothing of estimates over time should be
disabled. Default `FALSE`.}

\item{delta_tbar_prior_mean}{Prior mean for `delta_tbar`, the normal weight
on `theta_bar` in the previous period.  Default `0.5`.}

\item{delta_tbar_prior_sd}{Prior standard deviation for `delta_bar`. Default
`0.5`.}

\item{innov_sd_delta_scale}{Prior scale for `sd_innov_delta`, the Cauchy
innovation standard deviation of `nu_geo` and `delta_gamma`. Default `2.5`.}

\item{innov_sd_theta_scale}{Prior scale for `sd_innov_theta`, the Cauchy
innovation standard deviation of `gamma`, `xi`, and if `constant_item` is
`FALSE` the item difficulty `diff`. Default `2.5`.}

\item{...}{Further arguments passed to `\link[rstan]{stan}`.}
}
\value{
A `dgirtfit` object that extends `stanfit-class`. For details see
`\link{dgirtfit-class}`.
}
\description{
`dgirt` makes a call to `\link[rstan]{stan}` with the Stan code and data for
a DGIRT model. By default the call overrides the `stan` defaults for `pars`
to specify typical DGIRT parameters of interest and sets `iter_r` to `1L`.
}

