% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dgirtfit.r, R/methods-dgirtfit.r
\docType{class}
\name{dgirtfit-class}
\alias{dgirtfit}
\alias{dgirtfit-class}
\alias{extract,dgirtfit-method}
\alias{get_posterior_mean,dgirtfit-method}
\alias{show,dgirtfit-method}
\alias{summary,dgirtfit-method}
\title{Class \code{dgirtfit}: fitted DGIRT model}
\usage{
\S4method{show}{dgirtfit}(object = dgirtfit)

\S4method{summary}{dgirtfit}(object = dgirtfit, ...)

\S4method{extract}{dgirtfit}(object = dgirtfit, ...)

\S4method{get_posterior_mean}{dgirtfit}(object = dgirtfit,
  pars = "theta_bar", name = TRUE, ...)
}
\arguments{
\item{object}{A \code{dgirtfit}-class object.}

\item{...}{Further arguments to \code{\link{stanfit-class}} methods.}

\item{pars}{Selected parameter names.}

\item{name}{Whether to move descriptive rownames into columns.}
}
\description{
All \code{\link{rstan}} methods for superclass
\code{\link[rstan]{stanfit-class}} are available. Descriptive labels for
parameters on time periods, local geographic areas, and grouping variables
will be added to most output.

\code{show} method for \code{dgirtfit-class} objects

\code{summary} method for \code{dgirtfit-class} objects

\code{extract} method for \code{dgirtfit-class} objects

\code{get_posterior_mean} method for \code{dgirtfit-class} objects
}
\section{Slots}{

\describe{
\item{\code{dgirt_in}}{\code{\link{dgirtin-class}} data used to fit the model.}
}}
\examples{
data(toy_dgirtfit)
# summarize the fitted results
summary(toy_dgirtfit, pars = 'xi')

# get posterior means with a convenience function
get_posterior_mean(toy_dgirtfit, pars = 'xi')

# generally apply functions to posterior samples after warmup; n.b.
# `as.array` is iterations x chains x parameters so `MARGIN = 3` applies
# `FUN` over iterations and chains
apply(as.array(toy_dgirtfit, pars = 'xi'), 3, mean)

# access the posterior samples
as.array(toy_dgirtfit, pars = 'xi')  # samples after warmup
as.data.frame(toy_dgirtfit, pars = 'xi')  # all samples
extract(toy_dgirtfit, pars = 'theta_bar')  # all samples
toy_dgirtfit
summary(toy_dgirtfit)
extract(toy_dgirtfit)
get_posterior_mean(toy_dgirtfit)
}
\seealso{
\code{\link{stanfit-class}} \code{\link{dgirtin-class}}
}

