% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dgirtfit.r
\docType{class}
\name{dgirtfit-class}
\alias{dgirtfit}
\alias{dgirtfit-class}
\title{Class \code{dgirtfit}: a class for fitted DGIRT models}
\description{
Fitting a dgirt model results in a \code{dgirtfit} object that inherits from
\code{\link{rstan}}'s \code{\link[rstan]{stanfit-class}}. \code{rstan}
methods will be dispatched (only) if a \code{dgirtfit} method does
not exist.
}
\section{Slots}{

\describe{
\item{\code{dgirt_in}}{\code{\link{dgirtin-class}} data used to fit the model.}
}}
\examples{
data(toy_dgirtfit)
# summarize the fitted results
summary(toy_dgirtfit, pars = 'xi')

# get posterior means with a convenience function
get_posterior_mean(toy_dgirtfit, pars = 'theta_bar')

# generally apply functions to posterior samples after warmup; n.b.
# `as.array` is iterations x chains x parameters so `MARGIN = 3` applies
# `FUN` over iterations and chains
apply(as.array(toy_dgirtfit, pars = 'xi'), 3, mean)

# access the posterior samples
as.array(toy_dgirtfit, pars = 'theta_bar')
as.data.frame(toy_dgirtfit, pars = 'theta_bar')
extract(toy_dgirtfit, pars = 'theta_bar')
}
\seealso{
\code{\link{stanfit-class}} \code{\link{dgirtin-class}}
}

