% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-dgirtin.r, R/methods-dgirtin.r
\docType{methods}
\name{dgirtin-class}
\alias{dgirtin-class}
\alias{get_item_n,dgirtIn-method}
\alias{get_item_names,dgirtIn-method}
\alias{get_n,dgirtIn-method}
\title{Data for fitting a DGIRT Model.}
\usage{
\S4method{get_item_names}{dgirtIn}(x)

\S4method{get_n}{dgirtIn}(x, by = NULL, aggregate_name = NULL)

\S4method{get_item_n}{dgirtIn}(x, by = NULL, aggregate_data = FALSE)
}
\arguments{
\item{x}{An object of class `dgirtIn` as returned by `shape`.}

\item{by}{The name of a grouping variable.}

\item{aggregate_name}{If specified `get_n` will operate on the table passed
to `shape` as `aggregate_data` instead of on the individual data and count
nonmissingness in the given variable.}

\item{x}{An object of class `dgirtIn` as returned by `shape`.}
}
\value{
A list of item names.
}
\description{
Data for fitting a DGIRT Model.

Get Items Names in DGIRT Data.

Count Respondents in DGIRT Data.

Count Respondents to Items
}
\examples{
get_item_names(toy_dgirt_in)
# respondents count
get_n(toy_dgirt_in)

# respondent count by year
get_n(toy_dgirt_in, by = "year")

# respondent count by survey identifier
get_n(toy_dgirt_in, by = "source")

# specify `aggregate_name` to get counts of a variable in the aggregate data
get_n(toy_dgirt_in, aggregate_name = "race", by = "year")

get_item_n(toy_dgirt_in)
get_item_n(toy_dgirt_in, by = "year")
get_item_n(toy_dgirt_in, aggregate_data = TRUE)
get_item_n(toy_dgirt_in, by = "year", aggregate_data = TRUE)
}
\seealso{
`\link{get_item_n}, \link{get_item_names}`
}

