% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.r
\name{extract_dgirt}
\alias{extract_dgirt}
\title{Extract and name parameters}
\usage{
extract_dgirt(stan_output, stan_data)
}
\arguments{
\item{stan_output}{Return value of `dgirt`.}

\item{stan_data}{Return value of `wrangle`.}
}
\value{
Return value of `rstan::extract` with names attached to its elements.
}
\description{
`extract_dgirt` is a wrapper for rstan::extract that attaches
names to parameters according to the values of the data passed
to `dgirt`.
}

