% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modgirt.R
\name{modgirt}
\alias{modgirt}
\title{Fit a multidimensional, ordinal, dynamic group IRT model}
\usage{
modgirt(shaped_data, version = "2019_01_10_modgirt", model = NULL,
  evolving_alpha = FALSE, ...)
}
\arguments{
\item{shaped_data}{Data shaped for the model using
\code{\link{shape_modgirt}}.}

\item{version}{The name of the modgirt model to estimate, or the path to a
\code{.stan} file. Ignored if argument \code{model} is used.}

\item{model}{A Stan model object of class \code{stanmodel} to be used in
estimation. Specifying this argument avoids repeated model compilation. Note
that the Stan model object for a model fitted with \code{modgirt()} 
can be found in the the \code{stanfit} slot of the resulting
\code{modgirt_fit} object.}

\item{...}{Further arguments, passed to \code{\link[rstan]{stan}}.}
}
\value{
A \code{\link{modgirt_fit-class}} object.
}
\description{
Fit a multidimensional, ordinal, dynamic group IRT model
}
