% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dgirtfit.r
\docType{methods}
\name{poststratify}
\alias{poststratify}
\alias{poststratify,data.frame-method}
\alias{poststratify,dgirtfit-method}
\title{\code{poststratify}: reweight and aggregate estimates}
\usage{
poststratify(x, prop_name = "proportion", ...)

\S4method{poststratify}{data.frame}(x, target_data, strata_names,
  aggregated_names, estimate_names, prop_name = "proportion", keep = FALSE)

\S4method{poststratify}{dgirtfit}(x, target_data, strata_names,
  aggregated_names, estimate_names, prop_name = "proportion", keep = FALSE,
  pars = "theta_bar")
}
\arguments{
\item{x}{A \code{data.frame}.}

\item{prop_name}{Name of the column in \code{target_data} that gives
strata proportions.}

\item{...}{Additional arguments to methods.}

\item{target_data}{A table giving the proportions contributed to strata by
the interaction of \code{strata_names} and \code{aggregated_names}.}

\item{strata_names}{Names of variables whose interaction defines
population strata.}

\item{aggregated_names}{Names of variables to be aggregated over in
poststratification.}

\item{estimate_names}{Names of columns to be poststratified.}

\item{keep}{Whether to keep the original estimates and return them alongside
the poststratified estimates.}

\item{pars}{Selected parameter names.}
}
\value{
A table of poststratified estimates.
}
\description{
\code{poststratify}: reweight and aggregate estimates

Identifiers in the table of estimates to be poststratified should be given as
\code{strata_names} or \code{aggregated_names}. There will be a row in the
result for each interaction of the variables in \code{strata_names}
containing the values of \code{estimate_names} poststratified over the
variables in \code{aggregated_names}.

\code{poststratify} method for \code{dgirtfit}-class objects
}

