% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dgirtfit.r
\docType{methods}
\name{poststratify}
\alias{poststratify}
\alias{poststratify,data.frame-method}
\alias{poststratify,dgirtfit-method}
\title{\code{poststratify}: reweight and aggregate estimates}
\usage{
poststratify(x, prop_name = "proportion", ...)

\S4method{poststratify}{data.frame}(x = data.frame, target_data, strata_names,
  group_names, prop_name = "proportion", aggregate = FALSE,
  pars = "theta_bar")

\S4method{poststratify}{dgirtfit}(x = dgirtfit, target_data, strata_names,
  prop_name = "proportion", aggregate = FALSE, pars = "theta_bar")
}
\arguments{
\item{x}{A \code{dgirtfit}-class object.}

\item{prop_name}{The name of the column in \code{target_data} that gives
strata proportions.}

\item{...}{Additional arguments to methods.}

\item{target_data}{A table giving the proportions contributed to strata
populations by modeled groups.}

\item{strata_names}{The names of the columns in \code{x} and
\code{target_data} that define strata.}

\item{group_names}{The names of the columns in \code{x} and
\code{target_data} for grouping variables.}

\item{aggregate}{Whether to sum over multiple observations of strata and
grouping variables.}

\item{pars}{The names of model parameters of interest. Others will be
excluded.}
}
\value{
A table giving poststratified estimates for each stratum.
}
\description{
\code{poststratify}: reweight and aggregate estimates

\code{poststratify} method for \code{data.frame}s

\code{poststratify} method for \code{dgirtfit}-class objects
}

