% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dgirtfit-poststratify.r
\docType{methods}
\name{poststratify}
\alias{poststratify}
\alias{poststratify,data.frame-method}
\alias{poststratify,dgirtfit-method}
\title{\code{poststratify}: reweight and aggregate estimates}
\usage{
poststratify(x, target_data, strata_names, aggregated_names,
  prop_name = "proportion", ...)

\S4method{poststratify}{dgirtfit}(x, target_data, strata_names,
  aggregated_names, prop_name = "proportion", pars = "theta_bar")

\S4method{poststratify}{data.frame}(x, target_data, strata_names,
  aggregated_names, prop_name = "proportion", pars = "theta_bar")
}
\arguments{
\item{x}{A \code{data.frame}.}

\item{target_data}{A table giving the proportions contributed to strata by
the interaction of \code{strata_names} and \code{aggregated_names}.}

\item{strata_names}{Names of variables whose interaction defines
population strata.}

\item{aggregated_names}{Names of variables to be aggregated over in
poststratification.}

\item{prop_name}{Name of the column in \code{target_data} that gives
strata proportions.}

\item{...}{Additional arguments to methods.}

\item{pars}{Selected parameter names.}
}
\value{
A table of poststratified estimates.
}
\description{
\code{poststratify}: reweight and aggregate estimates

\code{poststratify} method for \code{dgirtfit}-class objects

Identifiers in the table of estimates to be poststratified should be given as
\code{strata_names} or \code{aggregated_names}. There will be a row in the
result for each interaction of the variables in \code{strata_names}
containing the values of \code{estimate_names} poststratified over the
variables in \code{aggregated_names}.
}
\examples{

data(toy_dgirtfit)

# the stratifying variables should uniquely identify proportions in the
# target data; to achieve this, sum over the other variables
targets <- aggregate(proportion ~ state + year + race, targets, sum)

# the dgirtfit method of poststratify takes a dgirtfit object, the target
# data, the names of variables that define population strata, and the  names
# of variables to be aggregated over
post <- poststratify(toy_dgirtfit, targets, c("state", "year"), "race")
}

