% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-dgirtfit-poststratify.r
\docType{methods}
\name{poststratify}
\alias{poststratify}
\alias{poststratify,data.frame-method}
\alias{poststratify,dgirtfit-method}
\title{\code{poststratify}: reweight and aggregate estimates}
\usage{
poststratify(x, target_data, strata_names, aggregated_names,
  prop_name = "proportion", single_issue = "F", ...)

\S4method{poststratify}{dgirtfit}(x, target_data, strata_names,
  aggregated_names, prop_name = "proportion", single_issue = "F",
  pars = "theta_bar")

\S4method{poststratify}{data.frame}(x, target_data, strata_names,
  aggregated_names, prop_name = "proportion", single_issue = "F",
  pars = "theta_bar")
}
\arguments{
\item{x}{A \code{data.frame} or \code{dgirtfit} object.}

\item{target_data}{A table giving the proportions contributed to strata by
the interaction of \code{strata_names} and \code{aggregated_names}.}

\item{strata_names}{Names of variables whose interaction defines
population strata.}

\item{aggregated_names}{Names of variables to be aggregated over in
poststratification.}

\item{prop_name}{Name of the column in \code{target_data} that gives
strata proportions.}

\item{single_issue}{Flag for whether DGO ran a single-issue manifest variable model. If "T", apply pnorm to convert results to response scale.}

\item{...}{Additional arguments to methods.}

\item{pars}{Selected parameter names.}
}
\value{
A table of poststratified estimates.
}
\description{
This function reweights and aggregates estimates from \code{dgirt} for strata
defined by modeled variables. The names of each of the model's time,
geographic, and demographic grouping variables can be given in either the
\code{strata_names} or \code{aggregated_names} argument. The result has
estimates for the strata indicated by the \code{strata_names} argument,
aggregated over the variables specified in \code{aggregated_names}.
\code{poststratify} requires a table given as \code{target_data} with
population proportions for the interaction of the variables given in
\code{strata_names} and \code{aggregated_names}.
}
\examples{

data(toy_dgirtfit)

# the stratifying variables should uniquely identify proportions in the
# target data; to achieve this, sum over the other variables
targets <- aggregate(proportion ~ state + year + race, targets, sum)

# the dgirtfit method of poststratify takes a dgirtfit object, the target
# data, the names of variables that define population strata, and the  names
# of variables to be aggregated over
post <- poststratify(toy_dgirtfit, targets, c("state", "year"), "race")
}

