% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratify.r, R/methods-dgirtfit.r
\docType{methods}
\name{post_generic}
\alias{post_generic}
\alias{poststratify,data.frame-method}
\alias{poststratify,dgirtfit-method}
\title{\code{poststratify}: post-stratify \code{dgirt} estimates}
\usage{
post_generic(x, target_data, strata_names, group_names,
  prop_name = "proportion", aggregate = FALSE, ...)

\S4method{poststratify}{data.frame}(x = data.frame, target_data, strata_names,
  group_names, prop_name = "proportion", aggregate = FALSE,
  pars = "theta_bar")

\S4method{poststratify}{dgirtfit}(x = dgirtfit, target_data, strata_names,
  prop_name = "proportion", aggregate = FALSE, pars = "theta_bar")
}
\arguments{
\item{x}{A \code{data.frame} or \code{dgirtfit} object.}

\item{target_data}{A table giving the proportions contributed to strata
populations by modeled groups.}

\item{strata_names}{The names of the columns in \code{x} and
\code{target_data} that define strata.}

\item{group_names}{The names of the columns in \code{x} and
\code{target_data} for grouping variables.}

\item{prop_name}{The name of the column in \code{target_data} that gives
strata proportions.}

\item{aggregate}{Whether to sum over multiple observations of strata and
grouping variables.}

\item{pars}{The names of model parameters of interest. Others will be
excluded.}

\item{x}{A \code{data.frame}.}

\item{target_data}{A table giving the proportions contributed to strata
populations by modeled groups.}

\item{group_names}{The names of the columns in \code{x} and
\code{target_data} for grouping variables.}

\item{strata_names}{The names of the columns in \code{x} and
\code{target_data} that define strata.}

\item{prop_name}{The name of the column in \code{target_data} that gives
strata proportions.}

\item{aggregate}{Whether to sum over multiple observations of strata and
grouping variables.}

\item{x}{A \code{dgirtfit}-class object.}

\item{target_data}{A table giving the proportions contributed to strata
populations by modeled groups.}

\item{group_names}{The names of the columns in \code{x} and
\code{target_data} for grouping variables.}

\item{strata_names}{The names of the columns in \code{x} and
\code{target_data} that define strata.}

\item{prop_name}{The name of the column in \code{target_data} that gives
strata proportions.}

\item{aggregate}{Whether to sum over multiple observations of strata and
grouping variables.}

\item{pars}{The names of model parameters of interest. Others will be
excluded.}
}
\value{
A table giving poststratified estimates for each stratum.

A table giving poststratified estimates for each stratum.

A table giving poststratified estimates for each stratum.
}
\description{
Given estimates for modeled groups and their incidence in population strata
of interest (typically defined in part geographically), \code{poststratify}
returns estimates for the strata. The estimate for each stratum is the mean
over the estimates for the groups it comprises, weighted by the proportion of
the stratum population the groups contribute.

\code{poststratify} method for \code{data.frame}s

\code{poststratify} method for \code{dgirtfit}-class objects
}
\examples{
#  dgirtfit method
poststratify(toy_dgirtfit, targets, strata_names = c("year", "state"), aggregate = TRUE)

# method for data.frame
x <- get_posterior_mean(toy_dgirtfit, pars = 'theta_bar')
poststratify(x, target_data = targets, group_names = "race", strata_names = c("year", "state") , aggregate = TRUE)

# this is a parameter indexed t
x <- get_posterior_mean(toy_dgirtfit, pars = 'xi')
poststratify(x, target_data = targets, group_names = "race", strata_names = c("year", "state") , aggregate = TRUE)
}

