% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.r
\name{shape}
\alias{shape}
\title{Shape Model Data.}
\usage{
shape(item_data, item_names, time_name, geo_name, group_names, weight_name,
  survey_name, modifier_data = NULL, target_data = NULL,
  aggregate_data = NULL, ...)
}
\arguments{
\item{item_data}{A table in which items appear in columns and each row
represents an individual's responses in some time period and local geographic
area.}

\item{item_names}{Individual item responses. These variables should be
integers or ordered factors in the data.}

\item{time_name}{A time variable with numeric values.}

\item{geo_name}{A geographic variable representing local areas.}

\item{group_names}{Discrete grouping variables, usually demographic. Using
numeric variables is allowed but not recommended.}

\item{weight_name}{A variable giving survey weights.}

\item{survey_name}{A survey identifier.}

\item{...}{Further arguments for more complex models, input data, and
preprocessing.}
}
\value{
An object of class `dgirtIn`, i.e., that expected by `\link{dgirt}`.
}
\description{
`shape` prepares data for modeling with `dgirt`.
}
\details{
`shape` takes four kinds of data. Only `item_data` is required. For each
other data type passed to `shape`, additional arguments are required, as
described in the sections below. Most arguments give the name or names of key
variables in the data; they end in `_name` or `_names` and should be
character vectors.  Some implement preprocessing and modeling choices.
}
\section{Modifier Data}{

These arguments are required to model hierarchical parameters. 
At the moment, modeling geographic parameters is supported
\describe{
  \item{`modifier_data`:}{Table giving characteristics of local geographic
  areas in time periods.}
  \item{`modifier_names`:}{Modifiers of geographic hierarchical parameters, e.g.
  median household income in each local-area and time-period combination.}
  \item{`t1_modifier_names`:}{Modifiers to be used instead of those in
  `modifier_names`, only in the first period.}
}
}

\section{Aggregate Data}{

Specifying `aggregate_data` requires no additional arguments; instead, we
make many assumptions about the data. This implementation is likely to change
in the future.
`aggregate_data` is expected to be a long table of trial and success counts
by group and item. Some variable names given for `item_data` are expected in
the table of aggregates: `group_names`, `geo_name`, and `time_name`. Three
fixed variable names are also expected in `aggregate_data`: `item` giving
item identifiers, `n_grp` giving adjusted counts of item-response trials, and
`s_grp` giving adjusted counts of item-response successes. The counts should
be adjusted consistently with the transformations applied to the individual
`item_data`.
}

\section{Preprocessing}{

If `target_data` is specified `shape` will adjust the weighting of groups
toward population targets via raking. This relies on an adaptation of
`\link[survey]{rake}`. The additional required arguments are
`target_proportion_name` and `strata_names`. The implementation will be more
flexible in the future, but at the moment the strata are defined additively
when more than one variable is given in `strata_names`. 

`shape` can restrict data row-wise in `item_data`, `modifier_data`, and
`aggregate_data` to that within specified time periods (`time_filter`) and
local geographic areas (`geo_filter`). Data can also be filtered for
sparsity, to keep items that appear in a minimum of time periods or surveys.
This is a column-wise operation. If both row-wise and column-wise
restrictions are specified, `shape` iterates over them until they leave the
data unchanged.

\describe{
  \item{`target_data`}{A table giving population proportions for groups by
  local geographic area and time period.}
  \item{target_proportion_name}{The variable giving population proportions
  for strata.}
  \item{strata_names}{Variables that define population strata.}
  \item{geo_filter}{A character vector giving values of the geographic
  variable. Defaults to observed values.}
  \item{time_filter}{A numeric vector giving possible values of the time
  variable. Observed and unobserved time periods can be given. Defaults to
  observed values.}
  \item{min_survey_filter}{An integer minimum of survey appearances for
  included items. Defaults to 1.}
  \item{min_t_filter}{An integer minimum of time period appearances for
  included items. Defaults to 1.}
}
}

\section{Modeling Choices}{

Optional. Most arguments like this one are now in the `dgirt` signature, but
`constant_item` affects the shape of the data. It may move to `dgirt` in the
future. 
\describe{
  \item{constant_item}{Whether item difficulty parameters should be constant
  over time. Default `TRUE`.}
}
}
\examples{
# model individual item responses 
data(state_opinion)
shaped_responses <- shape(state_opinion,
                          item_names = "Q_cces2006_minimumwage",
                          time_name = "year",
                          geo_name = "state",
                          group_names = "race",
                          weight_name = "weight",
                          survey_name = "source")
# summarize result)
summary(shaped_responses)
# check sparseness of data to be modeled
get_item_n(shaped_responses, by = "year")
}
\seealso{
get_n, get_item_n, get_item_names, dgirtin-class
}

