% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.r
\name{shape}
\alias{shape}
\title{Shape Model Data.}
\usage{
shape(item_data, ..., modifier_data = NULL, target_data = NULL,
  aggregate_data = NULL)
}
\arguments{
\item{item_data}{A table of individual item responses.}

\item{...}{Further arguments. See details below.}

\item{modifier_data}{A table giving characteristics of `item_data`
identifiers to be modeled as hierarchical parameters.}

\item{target_data}{A table of population proportions.}

\item{aggregate_data}{A table of group-level item-response aggregates.}
}
\value{
An object of the type expected by `\link{dgirt}`.
}
\description{
`shape` prepares data for modeling with `dgirt`.
}
\details{
`shape` takes four kinds of data. Only `item_data` is required. For each data
passed to `shape` through its `*_data` arguments, additional arguments are
required. These are described in the sections below. Most give the name or
names of key variables in the data; they end in `_name` or `_names` and
should be character vectors. A few implement preprocessing and modeling
choices.
}
\section{Item Data}{

These are required arguments if `item_data` is given.
\describe{
  \item{item_names}{Individual item responses. These variables should be
  integers or ordered factors in the data.}
  \item{group_names}{Discrete grouping variables, usually demographic. Using
  numeric variables is allowed but not recommended.}
  \item{geo_name}{A geographic variable representing local areas. In the
  data it should should be character or factor.}
  \item{time_name}{A time variable with numeric values.}
  \item{weight_name}{A variable giving survey weights.}
}
}

\section{Modifier Data}{

These are required arguments if `modifier_data` is given.
\describe{
  \item{modifier_names}{Modifiers of geographic hierarchical parameters, e.g.
  median household income in each local-area and time-period combination.}
  \item{t1_modifier_names}{Modifiers to be used instead of those in
  `modifier_names` only in the first period.}
}
}

\section{Aggregate Data}{

Specifying `aggregate_data` requires no additional arguments; instead, we
make many assumptions about the data. This implementation is likely to change
in the future. Variable names given for `item_data` are expected in the table
of aggregates: `item_names`, `group_names`, `geo_name`, and `time_name`. Two
fixed variable names are also expected in `aggregate_data`: `n_grp` giving
adjusted counts of item-response trials, and `s_grp` giving adjusted counts
of item-response successes. The counts should be adjusted consistently with
the transformations applied to the individual `item_data`.
}

\section{Preprocessing}{

If `target_data` is specified `shape` will adjust the weighting of groups
toward population targets via raking. This relies on an adaptation of
`\link[survey]{rake}`. The additional required arguments are
`target_proportion_name` and `strata_names`. The implementation will be more
flexible in the future, but at the moment the strata are defined additively
when more than one variable is given in `strata_names`. 

`shape` can restrict data row-wise in `item_data`, `modifier_data`, and
`aggregate_data` to that within specified time periods (`time_filter`) and
local geographic areas (`geo_filter`). Data can also be filtered for
sparsity, to keep items that appear in a minimum of time periods or surveys.
This is a column-wise operation. If both row-wise and column-wise
restrictions are specified, `shape` iterates over them until they leave the
data unchanged.

\describe{
  \item{target_proportion_name}{The variable giving population proportions
  for strata.}
  \item{strata_names}{Variables that define population strata.}
  \item{geo_filter}{A character vector giving values of the geographic
  variable. Defaults to observed values.}
  \item{time_filter}{A numeric vector giving possible values of the time
  variable. Observed and unobserved time periods can be given.}
  \item{min_survey_filter}{An integer minimum of survey appearances for
  included items. Defaults to 1 and requires `survey_name`.}
  \item{survey_name}{The variable that identifies surveys, for use with
  `min_survey_filter`.}
  \item{min_t_filter}{An integer minimum of time period appearances for
  included items.}
}
}

\section{Modeling Choices}{

Each of these arguments is optional. They may move to the `dgirt` signature
in the future.
\describe{
  \item{constant_item}{Whether item difficulty parameters should be constant
  over time. Default `TRUE`.}
  \item{separate_t}{Whether smoothing of estimates over time should be
  disabled. Default `FALSE`.}
  \item{delta_tbar_prior_mean}{Prior mean for `delta_tbar`, the normal weight
  on `theta_bar` in the previous period.  Default `0.5`.}
  \item{delta_tbar_prior_sd}{Prior standard deviation for `delta_bar`.
  Default `0.5`.}
  \item{innov_sd_delta_scale}{Prior scale for `sd_innov_delta`, the Cauchy
  innovation standard deviation of `nu_geo` and `delta_gamma`. Default `2.5`.}
  \item{innov_sd_theta_scale}{Prior scale for `sd_innov_theta`, the Cauchy
  innovation standard deviation of `gamma`, `xi`, and if `constant_item` is
  `FALSE` the item difficulty `diff`. Default `2.5`.}
}
}
\examples{
# model individual item responses 
data(state_opinion)
shaped_responses <- shape(state_opinion,
                          time_name = "year",
                          geo_name = "state",
                          item_names = "Q_cces2006_minimumwage",
                          group_names = "race",
                          survey_name = "source",
                          weight_name = "weight")
}

