% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.r
\name{shape}
\alias{shape}
\title{\code{shape}: prepare data for modeling with \code{dgirt}}
\usage{
shape(item_data, item_names, time_name, geo_name, group_names = NULL,
  weight_name, raking = NULL, survey_name, modifier_data = NULL,
  target_data = NULL, aggregate_data = NULL, ...)
}
\arguments{
\item{item_data}{A table in which items appear in columns and each row
represents an individual's responses in some time period and local geographic
area.}

\item{item_names}{Individual item responses. These variables should be
integers or ordered factors in the data.}

\item{time_name}{A time variable with numeric values.}

\item{geo_name}{A geographic variable representing local areas.}

\item{group_names}{Discrete grouping variables, usually demographic. Using
numeric variables is allowed but not recommended.}

\item{weight_name}{A variable giving survey weights.}

\item{survey_name}{A survey identifier.}

\item{modifier_data}{Table giving characteristics of local geographic areas
in time periods. See details below.}

\item{target_data}{A table giving population proportions for groups by local
geographic area and time period. See details below.}

\item{aggregate_data}{A table of trial and success counts by group and item.
See details below.}

\item{...}{Further arguments for more complex models, input data, and
preprocessing.}
}
\value{
An object of class \code{dgirtIn} expected by \code{\link{dgirt}}.
}
\description{
This function shapes various kinds of data for use in a dgirt model. Most
arguments give the name or names of key variables in the data; they end in
\code{_name} or \code{_names} and should be character vectors. Some others
implement preprocessing and modeling choices.
}
\section{Modifier Data}{

Geographic hierarchical parameters can be modeled with \code{modifier_data}.
These arguments are also required:
\describe{
  \item{\code{modifier_names}:}{Modifiers of geographic hierarchical
  parameters, e.g. median household income in each local-area and
  time-period combination.}
  \item{\code{t1_modifier_names}:}{Modifiers to be used instead of those in
  \code{modifier_names}, only in the first period.}
  \item{\code{standardize}:}{Whether to standardize hierarchical modifier
  data to be zero-mean and unit-variance for performance gains. For
  discussion see the Stan Language Reference section "Standardizing
  Predictors and Outputs."}
}
}

\section{Aggregate Data}{

Specifying \code{aggregate_data} requires no additional arguments; instead,
we make many assumptions about the data. This implementation is likely to
change in the future.

\code{aggregate_data} is expected to be a long table of trial and success
counts by group and item. Some variable names given for \code{item_data} are
expected in the table of aggregates: \code{group_names}, \code{geo_name}, and
\code{time_name}. Three fixed variable names are also expected in
\code{aggregate_data}: \code{item} giving item identifiers, \code{n_grp}
giving adjusted counts of item-response trials, and \code{s_grp} giving
adjusted counts of item-response successes. The counts should be adjusted
consistently with the transformations applied to the individual
\code{item_data}.
}

\section{Preprocessing}{

If \code{target_data} is specified \code{shape} will adjust the weighting of
groups toward population targets via raking. This relies on an adaptation of
\code{\link[survey]{rake}}. The additional required arguments are
\code{target_proportion_name} and \code{raking}. 

\code{shape} can restrict data row-wise in \code{item_data},
\code{modifier_data}, and \code{aggregate_data} to that within specified time
periods (\code{time_filter}) and local geographic areas (\code{geo_filter}).
Data can also be filtered for sparsity, to keep items that appear in a
minimum of time periods or surveys. This is a column-wise operation. If both
row-wise and column-wise restrictions are specified, \code{shape} iterates
over them until they leave the data unchanged.
}

\section{Target Data}{

\describe{
  \item{target_proportion_name}{The variable giving population proportions
  for strata.}
  \item{raking}{A formula or list of formulas specifying the variables on
  which to rake.}
  \item{geo_filter}{A character vector giving values of the geographic
  variable. Defaults to observed values.}
  \item{time_filter}{A numeric vector giving possible values of the time
  variable. Observed and unobserved time periods can be given. Defaults to
  observed values.}
  \item{min_survey_filter}{An integer minimum of survey appearances for
  included items. Defaults to 1.}
  \item{min_t_filter}{An integer minimum of time period appearances for
  included items. Defaults to 1.}
}
}

\section{Modeling Choices}{

Optional. Most arguments like this one are now in the \code{dgirt} signature,
but \code{constant_item} affects the shape of the data. It may move to
\code{dgirt} in the future.
\describe{
  \item{constant_item}{Whether item difficulty parameters should be constant
  over time. Default \code{TRUE}.}
}
}
\examples{
# model individual item responses
data(opinion)
shaped_responses <- shape(opinion,
                          item_names = "Q_cces2006_gaymarriageamendment",
                          time_name = "year",
                          geo_name = "state",
                          group_names = "race",
                          weight_name = "weight",
                          survey_name = "source")
# summarize result)
summary(shaped_responses)

# check sparseness of data to be modeled
get_item_n(shaped_responses, by = "year")

}
\seealso{
\code{\link{dgirtin-class}} \code{\link{dgirtfit-class}}
}

