% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdata-state_opinion.r
\docType{data}
\name{state_opinion}
\alias{state_opinion}
\title{State-level policy item responses}
\format{A `data.frame` with 295,248 rows and 33 columns.}
\usage{
state_opinion
}
\description{
A table of item responses given by survey respondents, and their
characteristics. Data are from the Cooperative Congressional Election Study
(CCES), 2006-2014.
}
\examples{
state_opinion
}
\seealso{
\url{http://projects.iq.harvard.edu/cces/data}
}

