% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-states.r
\docType{data}
\name{states}
\alias{states}
\title{State-level policy item responses}
\format{A \code{dplyr} \code{tbl_df} with 295,248 rows and 33 columns.}
\usage{
states
}
\description{
A dataset of survey item responses and characteristics of respondents.
Data are from the Cooperative Congressional Election Study (CCES) 2006-2014.
`Q_` prepends the names of item response variables and `D_` those of respondent characteristics.
}
\examples{
states
}

