% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-targets.r
\docType{data}
\name{targets}
\alias{targets}
\title{\code{dgirt} example data: U.S. population targets}
\format{A \code{data.frame} with ~1.6M observations of 10 variables.}
\usage{
targets
}
\description{
A table giving U.S. population proportions by year for combinations of
demographic variables. For use with the \code{target_data} argument of
\code{\link{shape}}. Data are from the U.S. Census. Proportions sum to to 1
within years.
}
\examples{
targets
}
\seealso{
\url{http://www.census.gov/}
}

