% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-targets.r
\docType{data}
\name{targets}
\alias{targets}
\title{Yearly U.S. Population Proportions.}
\format{A `data.frame` with 18,360 observations of 6 variables.}
\usage{
targets
}
\description{
A table giving U.S. population proportions by year for combinations of
demographic variables. Data are from the U.S. Census.
When using this table to adjust survey weights via the `target_data`
arguments of `shape`, note that its proportions sum to to 1 within years.
}
\examples{
targets
}
\seealso{
\url{http://www.census.gov/}
}

