% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-toy_dgirt_in.r
\docType{data}
\name{toy_dgirt_in}
\alias{toy_dgirt_in}
\title{A Minimal Example of the dgirtIn Class.}
\format{A `dgirtIn` object.}
\usage{
toy_dgirt_in
}
\description{
`\link{shape}` returns a `dgirtIn`-class object used with `\link{dgirt}` for
DGIRT modeling. `toy_dgirt_in` is a minimal `dgirtIn` object, mostly for
use in development.
}
\examples{
# created as follows
data(state_opinion)
toy_dgirt_in <- shape(state_opinion,
                 time_name = "year",
                 item_names = c("Q_cces2006_minimumwage",
                                "Q_cces2006_abortion"),
                 geo_name = "state",
                 group_names = "race",
                 survey_name = "source",
                 weight_name = "weight")
}

