% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-toy_dgirt_in.r
\docType{data}
\name{toy_dgirt_in}
\alias{toy_dgirt_in}
\title{Class \code{dgirtIn}: a minimal example object}
\format{A \code{\link{dgirtin-class}} object.}
\usage{
toy_dgirt_in
}
\description{
\code{\link{shape}} returns a \code{\link{dgirtin-class}} object used with
\code{\link{dgirt}} for DGIRT modeling. \code{toy_dgirt_in} is a minimal
\code{\link{dgirtin-class}} object, mostly for use in development.
}
\examples{
toy_dgirt_in
}

