% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_dgirt_in.r
\docType{data}
\name{toy_dgirt_in}
\alias{toy_dgirt_in}
\title{Class \code{dgirtIn}: a minimal example object}
\format{A \code{\link[dgo]{dgirtin-class}} object.}
\usage{
toy_dgirt_in
}
\description{
\code{\link[dgo]{shape}} returns a \code{\link[dgo]{dgirtin-class}} object
used with \code{\link[dgo]{dgirt}} for DGIRT modeling. \code{toy_dgirt_in} is
a minimal \code{\link[dgo]{dgirtin-class}} object for use in examples.
}
\examples{
toy_dgirt_in
}

