% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-toy_dgirtfit.r
\docType{data}
\name{toy_dgirtfit}
\alias{toy_dgirtfit}
\title{A Minimal Example of the `dgirtfit` Class.}
\format{A `dgirtfit` object.}
\usage{
toy_dgirtfit
}
\description{
`\link{dgirt}` returns a `dgirtfit`-class object that extends the `stanfit`
class from `rstan`.  `toy_dgirtfit` is a minimal `dgirtfit` object, mostly
for use in development.
}
\examples{
toy_dgirtfit
}

