% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-toy_dgirtfit.r
\docType{data}
\name{toy_dgirtfit}
\alias{toy_dgirtfit}
\title{Class \code{dgirtfit}: a minimal example object}
\format{A \code{\link{dgirtfit-class}} object.}
\usage{
toy_dgirtfit
}
\description{
\code{\link{dgirt}} returns a \code{\link{dgirtfit}}-class object that
extends \code{\link[rstan]{stanfit-class}} from \code{\link{rstan}}.
\code{toy_dgirtfit} is a minimal \code{\link{dgirtfit}} object, mostly for
use in development.
}
\examples{
toy_dgirtfit
}

