% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_dgirtfit.r
\docType{data}
\name{toy_dgirtfit}
\alias{toy_dgirtfit}
\title{Class \code{dgirtfit}: a minimal example object}
\format{A \code{\link[dgo]{dgirtfit-class}} object.}
\usage{
toy_dgirtfit
}
\description{
\code{\link[dgo]{dgirt}} returns a \code{\link[dgo]{dgirtfit}}-class object
that extends \code{\link[rstan]{stanfit-class}}. \code{toy_dgirtfit} is a
minimal \code{\link[dgo]{dgirtfit}} object for use in examples.
}
\examples{
toy_dgirtfit
}

